refactor: extract ReadingPaneView to its own file (check_shell_chrome, 1/2)
ContentView+ViewBuilders.swift was 1320 lines, over the guardrail's 1000-line
shell limit. Extracted a COHESIVE slice, not a line-count cut: ReadingPaneView was
already a self-contained `private struct` — the whole reading surface with its own
pin state, zoom/notes/page/knowledge tabs — sitting at the top of a file whose job
is ContentView's view builders. It never belonged there.

- New: Views/Library/Reading/ReadingPaneView.swift (564 lines), beside its
  siblings (PageContentPane, PDFThumbnailView, ImmersiveReaderView). The name says
  exactly what is in it.
- ContentView+ViewBuilders.swift: 1320 → 757 lines.
- `private struct` → `struct`: it is now referenced across files. Every stored
  property the one call site passes (liveDocument, liveActivePageNumber,
  livePageCount, scrollSync, onPageSelected, onClose) is internal, so the
  synthesized memberwise init stays internal and the call site compiles.
- project.pbxproj: 4 entries (build file / file ref / group / sources). The app
  target is a non-synchronized group — a missed membership is a "cannot find type"
  build failure. plutil -lint passes.

PURE extraction — verified by diffing the moved block and the remainder against
HEAD: byte-identical apart from the access modifier and two blank lines. Same
views, same modifiers, same order. No new computed properties and no new body, so
nothing new for the type-checker to chew on (the LibraryWindow.body timeout hazard
is not touched — this only makes the type-checker's job smaller).

swiftlint clean on both files. check_shell_chrome still red on ContentView.swift
(1022) — that is commit 2/2, kept separate so a build failure bisects cleanly.

(cherry picked from commit 8b0c03812bc64f96123367ef7f3570e1226b640a)
